Enhanced performace by only using regex
