--- src/app/directives/text-find.directive.spec.ts
@@ -1,8 +1,0 @@
-import { TextFindDirective } from './text-find.directive';
-
-describe('TextFindDirective', () => {
-  it('should create an instance', () => {
-    const directive = new TextFindDirective();
-    expect(directive).toBeTruthy();
-  });
-});
