+++ src/app/document/document.component.html
@@ -1,30 +1,32 @@
-<ng-container *ngIf="loading">
-  <app-loader></app-loader>
-  <p class="primary-text text-center">Building document...</p>
-</ng-container>
+<div #documentEl >
+  <ng-container *ngIf="loading">
+    <app-loader></app-loader>
+    <p class="primary-text text-center">Building document...</p>
+  </ng-container>
 
-<ng-container *ngIf="error">
-  <app-error [message]="'Sorry there was a problem loading this document'"></app-error>
-</ng-container>
+  <ng-container *ngIf="error">
+    <app-error [message]="'Sorry there was a problem loading this document'"></app-error>
+  </ng-container>
 
-<ng-container *ngIf="document" >
-  <div appTextFind *ngFor="let contentObj of document.content; let pIndex = index">
-    <p *ngIf="contentObj.type === 'paragraph'">
-      <ng-container *ngFor="let content of contentObj.content; let spanIndex = index ">
-        <ng-container *ngIf="content.type === 'span'">
-          <span
-          [ngClass]="{
-            'font-bold': content.attrs.bold || content.attrs.textCSS && content.attrs.textCSS.fontWeight === 'bold' || false,
-            'font-italic': content.attrs.italic || content.attrs.textCSS && content.attrs.textCSS.fontStyle === 'italic' || false,
-            'primary-font-family': content.attrs.font === 'Times New Roman' || content.attrs && content.attrs.textCSS && content.attrs.textCSS.fontFamily === 'Times New Roman' || false,
-            'text-underline': content.attrs.textCSS && content.attrs.textCSS.textDecoration && content.attrs.textCSS.textDecoration === 'underline' || false,
-            'text-decoration-solid': content.attrs.textCSS && content.attrs.textCSS.textDecorationStyle && content.attrs.textCSS.textDecorationStyle === 'solid' || false }">
-            <ng-container *ngFor="let item of content.content; let contentIndex = index">
-              <span class="doc-text" [id]="pIndex +'-'+ spanIndex +'-' + contentIndex">{{ item.text }}</span>
-            </ng-container>
-          </span>
+  <ng-container *ngIf="document" >
+    <div *ngFor="let contentObj of document.content; let pIndex = index">
+      <p *ngIf="contentObj.type === 'paragraph'">
+        <ng-container *ngFor="let content of contentObj.content; let spanIndex = index ">
+          <ng-container *ngIf="content.type === 'span'">
+            <span
+            [ngClass]="{
+              'font-bold': content.attrs.bold || content.attrs.textCSS && content.attrs.textCSS.fontWeight === 'bold' || false,
+              'font-italic': content.attrs.italic || content.attrs.textCSS && content.attrs.textCSS.fontStyle === 'italic' || false,
+              'primary-font-family': content.attrs.font === 'Times New Roman' || content.attrs && content.attrs.textCSS && content.attrs.textCSS.fontFamily === 'Times New Roman' || false,
+              'text-underline': content.attrs.textCSS && content.attrs.textCSS.textDecoration && content.attrs.textCSS.textDecoration === 'underline' || false,
+              'text-decoration-solid': content.attrs.textCSS && content.attrs.textCSS.textDecorationStyle && content.attrs.textCSS.textDecorationStyle === 'solid' || false }">
+              <ng-container *ngFor="let item of content.content; let contentIndex = index">
+                <span class="doc-text" [id]="pIndex +'-'+ spanIndex +'-' + contentIndex">{{ item.text }}</span>
+              </ng-container>
+            </span>
+          </ng-container>
         </ng-container>
-      </ng-container>
-    </p>
-  </div>
-</ng-container>
+      </p>
+    </div>
+  </ng-container>
+</div>
