+++ src/styles/base/_forms.scss
@@ -20,6 +20,7 @@ button {
   color: #fff;
   border-radius: 15px;
   padding: 0 15px;
+  cursor: pointer;
 
   &:active,
   &:focus {
